requirements.txt
streamlit==1.22.0
python-docx==0.8.11
